# docs/runtime_api_openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Governance Atlas Runtime",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/lineage/column-trace": {
      "get": {
        "description": "Query-param variant so asset_fqn (which contains dots) doesn't\nhave to collide with FastAPI's dot-unfriendly path parameter.",
        "operationId": "api_column_lineage_trace_query_api_lineage_column_trace_get",
        "parameters": [
          {
            "in": "query",
            "name": "asset_fqn",
            "required": false,
            "schema": {
              "default": "",
              "title": "Asset Fqn",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "column_name",
            "required": false,
            "schema": {
              "default": "",
              "title": "Column Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "upstream",
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 2,
              "title": "Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Column Lineage Trace Query",
        "tags": [
          "lineage"
        ]
      }
    },
    "/api/lineage/columns/{asset_fqn}/{column_name}/trace": {
      "get": {
        "description": "Phase 9 \u2014 multi-hop column lineage. Walks system.access.column_lineage\nrecursively, bounded by depth/node/fanout caps to fail closed against\nrunaway fan-out.",
        "operationId": "api_column_lineage_trace_api_lineage_columns__asset_fqn___column_name__trace_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_fqn",
            "required": true,
            "schema": {
              "title": "Asset Fqn",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "column_name",
            "required": true,
            "schema": {
              "title": "Column Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "upstream",
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 2,
              "title": "Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Column Lineage Trace",
        "tags": [
          "lineage"
        ]
      }
    },
    "/api/lineage/{asset_fqn}": {
      "get": {
        "operationId": "api_lineage_api_lineage__asset_fqn__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_fqn",
            "required": true,
            "schema": {
              "title": "Asset Fqn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Lineage",
        "tags": [
          "lineage"
        ]
      }
    },
...
  }
}